feat: Move scroll animation logic to a dedicated Client Component (refs #21)
